flowise_duplicate_tickets/requirements.txt
pandas==2.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
gensim==4.3.3
flask==3.0.3
requests==2.32.3
scikit-learn==1.5.0
flask-cors==4.0.1
